Address RESP cache review comments
Move the regression into RespTests, shorten the cache guard comment, and apply repository formatting.
