+++ test/standalone/Garnet.test/RespTests.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
+using System.Net.Sockets;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
@@ -5473,6 +5474,65 @@ namespace Garnet.test
             var mainDB = mainConnection.GetDatabase(0);
 
             Assert.Throws<RedisServerException>(() => mainDB.Execute("CLIENT", "UNBLOCK", 123, "INVALID"));
+        }
+
+        /// <summary>
+        /// Verifies that the command cache does not truncate argument counts.
+        /// </summary>
+        [Test]
+        [TestCase(255)]
+        [TestCase(256)]
+        [TestCase(259)]
+        [TestCase(600)]
+        public void CachedCommandDoesNotTruncateArgumentCount(int argCount)
+        {
+            const string InjectedCommand = "*1\r\n$4\r\nPING\r\n";
+
+            using var socket = new Socket(TestUtils.EndPoint.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
+            socket.Connect(TestUtils.EndPoint);
+
+            // Prime the cache with a benign frame that shares the exact 16-byte header prefix.
+            socket.Send(BuildDelCommand(argCount, injectAt: -1, InjectedCommand));
+            ClassicAssert.AreEqual(":0\r\n", ReadSocketResponses(socket));
+
+            // Place a complete RESP command where a truncated argument count would resume parsing.
+            socket.Send(BuildDelCommand(argCount, injectAt: argCount & 0xFF, InjectedCommand));
+            ClassicAssert.AreEqual(":0\r\n", ReadSocketResponses(socket));
+        }
+
+        private static byte[] BuildDelCommand(int argCount, int injectAt, string injectedCommand)
+        {
+            var sb = new StringBuilder();
+            sb.Append($"*{argCount + 1}\r\n$3\r\nDEL\r\n");
+            for (var i = 0; i < argCount; i++)
+            {
+                var arg = i == injectAt ? injectedCommand : $"k{i}";
+                sb.Append($"${arg.Length}\r\n{arg}\r\n");
+            }
+
+            return Encoding.ASCII.GetBytes(sb.ToString());
+        }
+
+        private static string ReadSocketResponses(Socket socket)
+        {
+            var sb = new StringBuilder();
+            var buffer = new byte[8192];
+            var deadline = DateTime.UtcNow.AddSeconds(5);
+
+            while (DateTime.UtcNow < deadline)
+            {
+                if (socket.Available == 0)
+                {
+                    Thread.Sleep(10);
+                    continue;
+                }
+
+                var read = socket.Receive(buffer);
+                sb.Append(Encoding.ASCII.GetString(buffer, 0, read));
+                deadline = DateTime.UtcNow.AddMilliseconds(500);
+            }
+
+            return sb.ToString();
         }
 
         [Test]
